+++ libs/mock-block-dock/src/datastore/use-mock-datastore/non-temporal.ts
@@ -45,6 +45,7 @@ export type MockDatastore = {
       | "deleteEntityType"
       | "createPropertyType"
       | "updatePropertyType"
+      | "requestLinkedQuery"
     >
   >;
 };