+++ libs/block-template-html/CHANGELOG.md
@@ -1,5 +1,12 @@
 # block-template-html
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - block-scripts@0.3.2
+
 ## 0.3.2
 
 ### Patch Changes